Reword "Basic example" sentence in index.rst.
The original sentence started, "To sdist-package, install and test your project..."  Although it strictly makes sense, using sdist-package as a transitive verb was a real stumbling block for me here.  It looks like a sentence that got broken in an edit.  Like "Type <something> to sdist-package," or "In order to sdist-package, first install and test..."

My proposed replacement is,  "Now, to package, install and test your project..."

I think people who recognize "sdist-package" as a verb will see it as implicit in "package", while idiots like me won't be underserved at this point in the Basic Example.  But another rephrasing that does mention sdist would be,

"To create an sdist package for your project, and install and test it..."